fix(session): open duplicated tab instantly and unify duplicate entry points
- In the shared duplicate routine, create the new tab right after the session
  is bound and BEFORE SessionStart so it appears immediately. Previously the tab
  was created only after SessionStart, and waitForTerminalSize ran before the
  new terminal existed (blocking ~1.5s), so the tab showed up ~1s+ late.
- Consolidate the keyboard-shortcut duplicateSession and the context-menu
  duplicate into one shared useDuplicateSession() routine used by both.

Closes #584
